# docs/swagger.yaml
basePath: /v1
definitions:
  dto.AccountLoginReq:
    properties:
      paswd:
        maxLength: 255
        minLength: 8
        type: string
      uname:
        maxLength: 16
        type: string
    required:
    - paswd
    - uname
    type: object
  dto.AccountLoginResp:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  dto.AccountProfileResp:
    properties:
      age:
        type: integer
      fullname:
        type: string
      gender:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.PersonDetailResp:
    properties:
      age:
        type: integer
      fullname:
        type: string
      gender:
        type: string
      id:
        type: integer
    type: object
  dto.PersonUpdateReq:
    properties:
      birth_date:
        type: string
      fullname:
        maxLength: 56
        minLength: 4
        type: string
      gender:
        enum:
        - m
        - f
        type: string
    required:
    - birth_date
    - fullname
    - gender
    type: object
  dto.SuccessResponse-any:
    properties:
      data: {}
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SuccessResponse-array_dto_PersonDetailResp:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.PersonDetailResp'
        type: array
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SuccessResponse-dto_AccountLoginResp:
    properties:
      data:
        $ref: '#/definitions/dto.AccountLoginResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SuccessResponse-dto_AccountProfileResp:
    properties:
      data:
        $ref: '#/definitions/dto.AccountProfileResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SuccessResponse-dto_PersonDetailResp:
    properties:
      data:
        $ref: '#/definitions/dto.PersonDetailResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:3000
info:
  contact:
    email: mark.p.e.muhammad@gmail.com
    name: Mark Muhammad
  description: This is a boilerplate project, please update accordingly.
  license:
    name: MIT
  title: Base API Service
  version: "1.0"
paths:
  /accounts:
    get:
      description: Get profile of logged-in account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AccountProfileResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get account's profile
  /accounts/login:
    post:
      consumes:
      - application/json
      description: Account login using username & password combination.
      parameters:
      - description: Credential
        in: body
        name: cred
        required: true
        schema:
          $ref: '#/definitions/dto.AccountLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AccountLoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Account login
  /persons:
    get:
      description: Get a list of person.
      parameters:
      - description: Person's name
        in: query
        name: q
        type: string
      - description: Data offset
        in: query
        name: s
        type: integer
      - description: Data limit
        in: query
        name: l
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-array_dto_PersonDetailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a list of person
  /persons/{id}:
    get:
      description: Get a person's detail.
      parameters:
      - description: Person's ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_PersonDetailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a person's detail
    put:
      consumes:
      - application/json
      description: Update a person's detail.
      parameters:
      - description: Person's ID
        in: path
        name: id
        required: true
        type: integer
      - description: Person's detail
        in: body
        name: detail
        required: true
        schema:
          $ref: '#/definitions/dto.PersonUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a person's detail
securityDefinitions:
  BearerAuth:
    description: Bearer auth containing JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"